+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is influenced by http://keepachangelog.com/.
 - Upgrade Rails to v5.0.0.1
 - Upgrade UI kit to v1.7.5
 - Improvements to Category landing page layout including collapsing regions
+- Remove references to content analysis service
 
 ### Fixed
 - Fixed leaking email addresses on password reset and resend email confirmation pages